extract styles
